Add support for Magisk Alpha2

+++ edxp-core/template_override/uninstall.sh
@@ -1,3 +1,19 @@
 #!/sbin/sh
 
-rm -rf /data/misc/riru/modules/edxp
+MODDIR=${0%/*}
+VARIANT="YAHFA"
+REMOVE=false
+
+[[ "$(echo ${MODDIR} | grep sandhook)" != "" ]] && VARIANT="SandHook"
+
+if [[ "${VARIANT}"=="SandHook" ]]; then
+    [[ -f "${MODDIR}/../riru_edxposed/module.prop" ]] || REMOVE=true
+else
+	[[ -f "${MODDIR}/../riru_edxposed_sandhook/module.prop" ]] || REMOVE=true
+fi
+
+if [[ "${REMOVE}" = true ]]; then
+    rm -rf /data/misc/riru/modules/edxp
+fi
+
+
